Moved all of the resource specific utils to their resource files
