+++ ec2/tests/test_ec2_elasticip.py
@@ -299,3 +299,11 @@ class TestElasticIP(testtools.TestCase):
         self.assertNotIn(
             'aws_resource_id', ctx.instance.runtime_properties.keys())
         self.assertIsNotNone(ec2_client.get_all_addresses(address.public_ip))
+
+    @mock_ec2
+    def test_get_all_addresses_bad(self):
+
+        ctx = self.mock_ctx('test_get_all_addresses_bad')
+        output = elasticip._get_all_addresses(
+            address='127.0.0.1', ctx=ctx)
+        self.assertIsNone(output)
