+++ ec2/tests/test_ec2_utils.py
@@ -20,14 +20,13 @@ import testtools
 from moto import mock_ec2
 
 # Cloudify Imports is imported and used in operations
-from ec2 import connection
 from ec2 import utils
+from ec2 import securitygroup
 from cloudify.mocks import MockCloudifyContext
 from cloudify.exceptions import NonRecoverableError
 
 TEST_AMI_IMAGE_ID = 'ami-e214778a'
 TEST_INSTANCE_TYPE = 't1.micro'
-FQDN = '((?:[a-z][a-z\\.\\d\\-]+)\\.(?:[a-z][a-z\\-]+))(?![\\w\\.])'
 
 
 class TestUtils(testtools.TestCase):
@@ -53,92 +52,10 @@ class TestUtils(testtools.TestCase):
 
         return ctx
 
-    def test_get_instance_state(self):
-        """ this tests that get instance state returns
-        running for a running instance
-        """
-
-        ctx = self.mock_ctx('test_get_instance_state')
-        with mock_ec2():
-            ec2_client = connection.EC2ConnectionClient().client()
-            reservation = ec2_client.run_instances(
-                TEST_AMI_IMAGE_ID, instance_type=TEST_INSTANCE_TYPE)
-            ctx.instance.runtime_properties['aws_resource_id'] = \
-                reservation.instances[0].id
-            instance_state = utils.get_instance_state(ctx=ctx)
-            self.assertEqual(instance_state, 16)
-
-    def test_no_instance_get_instance_from_id(self):
-        """ this tests that a NonRecoverableError is thrown
-        when a nonexisting instance_id is provided to the
-        get_instance_from_id function
-        """
-
-        ctx = self.mock_ctx('test_no_instance_get_instance_from_id')
-
-        with mock_ec2():
-
-            instance_id = 'bad_id'
-            output = utils.get_instance_from_id(instance_id, ctx=ctx)
-            self.assertIsNone(output)
-
-    def test_get_private_dns_name(self):
-
-        ctx = self.mock_ctx('test_get_private_dns_name')
-
-        with mock_ec2():
-            ec2_client = connection.EC2ConnectionClient().client()
-            reservation = ec2_client.run_instances(
-                TEST_AMI_IMAGE_ID, instance_type=TEST_INSTANCE_TYPE)
-            ctx.instance.runtime_properties['aws_resource_id'] = \
-                reservation.instances[0].id
-            property_name = 'private_dns_name'
-            dns_name = utils.get_instance_attribute(property_name, ctx=ctx)
-            self.assertRegexpMatches(dns_name, FQDN)
-
-    def test_get_key_pair_by_id(self):
-        with mock_ec2():
-            ec2_client = connection.EC2ConnectionClient().client()
-            kp = ec2_client.create_key_pair('test_get_key_pair_by_id_bad_id')
-            output = utils.get_key_pair_by_id(kp.name)
-            self.assertEqual(output.name, kp.name)
-
-    def test_get_security_group_from_name(self):
-
-        ctx = self.mock_ctx('test_get_security_group_from_name')
-
-        with mock_ec2():
-            ec2_client = connection.EC2ConnectionClient().client()
-            group = \
-                ec2_client.create_security_group('test_get_'
-                                                 'security_group_from_name',
-                                                 'this is test')
-            output = utils.get_security_group_from_name(group.id, ctx=ctx)
-            self.assertEqual(group.id, output.id)
-
-    @mock_ec2
-    def test_get_all_groups_deleted(self):
-
-        ctx = self.mock_ctx('test_get_all_groups_deleted')
-        ec2_client = connection.EC2ConnectionClient().client()
-        group = ec2_client.create_security_group('test_get_all_groups_deleted',
-                                                 'this is test')
-        output = utils.get_all_security_groups(
-            list_of_group_ids=group.id, ctx=ctx)
-        self.assertEqual(output[0].id, group.id)
-
-    @mock_ec2
-    def test_get_all_addresses_bad(self):
-
-        ctx = self.mock_ctx('test_get_all_addresses_bad')
-        output = utils.get_all_addresses(
-            address='127.0.0.1', ctx=ctx)
-        self.assertIsNone(output)
-
     @mock_ec2
     def test_log_available_resources(self):
         ctx = self.mock_ctx('test_log_available_resources')
-        groups = utils.get_all_security_groups(ctx=ctx)
+        groups = securitygroup._get_all_security_groups(ctx=ctx)
         utils.log_available_resources(groups, ctx=ctx)
 
     @mock_ec2
@@ -150,11 +67,3 @@ class TestUtils(testtools.TestCase):
             'image_id', ctx=ctx)
         self.assertIn(
             'is a required input', ex.message)
-
-    @mock_ec2
-    def test_get_all_instances_bad_id(self):
-
-        ctx = self.mock_ctx('test_get_all_instances_bad_id')
-        output = utils.get_all_instances(
-            list_of_instance_ids='test_get_all_instances_bad_id', ctx=ctx)
-        self.assertIsNone(output)
